extract caching logic to JwtCachingService
